fix: organizer column and moved selectColor function from columns to separate file in utils
